refactor(): updated creditor payments view

+++ src/views/creditorPayments/controller/credPaymentsController.jsx
@@ -10,19 +10,8 @@ import { backendAuthApi } from 'src/axios/instance/backend-axios-instance';
 import { BACKEND_API } from 'src/axios/constant/backend-api';
 import responseUtil from 'src/utils/responseUtil';
 import commonUtil from 'src/utils/common-util';
-import { PAYMENT_STATUS } from 'src/constants/commonConstants';
 
-const validationInvoiceUpdate = Yup.object().shape({
-  credInvoiceNo: Yup.string().required('Invoice no is required'),
-  credInvoiceDate: Yup.string().required('Invoiced Date is required'),
-  credInvoiceAmount: Yup.number().required().min(0, 'Invoice amount is invalid'),
-  credInvoicePaidDate: Yup.string().required('Invoice paid date is required'),
-  credInvoiceStatus: Yup.string()
-    .required('Status required')
-    .oneOf([PAYMENT_STATUS.PAID, PAYMENT_STATUS.NOTPAID], 'Invalid Status'),
-});
-
-const validationInvoicePayment = Yup.object().shape({
+const validationSchema = Yup.object().shape({
   invoiceNo: Yup.string().required('Invoice no is required'),
   invoiceAmount: Yup.number().required().min(0, 'Invoice amount is invalid'),
   invoiceCreatedAt: Yup.string().required('Invoice created date is required'),
@@ -30,15 +19,15 @@ const validationInvoicePayment = Yup.object().shape({
 
 const CredPaymentsController = () => {
   const headerLabels = [
-    '',
+    'Main Invoice',
+    'Sub Invoice',
     'Creditor',
-    'Invoice No',
-    'Amount',
+    'Invoice Amount',
     'Balance Amount',
+    'Paid Amount',
     'Invoiced Date',
-    'Due Date',
-    'Status',
     'Paid Date',
+    'Due Date',
   ];
 
   const sourceToken = axios.CancelToken.source();
@@ -53,37 +42,21 @@ const CredPaymentsController = () => {
   const [count, setCount] = useState(0);
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
-  const [isOpenAdd, setIsOpenAdd] = useState(false);
   const [isOpenUpdate, setIsOpenUpdate] = useState(false);
   const [isOpenDelete, setIsOpenDelete] = useState(false);
 
   const [isLoading, setIsLoading] = useState(false);
   const [isLoadingTotal, setIsLoadingTotal] = useState(true);
-  const [isLoadingAddPayment, setIsLoadingAddPayment] = useState(false);
   const [isLoadingUpdate, setIsLoadingUpdate] = useState(false);
   const [isLoadingDelete, setIsLoadingDelete] = useState(false);
 
   const formik = useFormik({
     initialValues: {
-      credInvoiceNo: '',
-      credInvoiceDate: null,
-      credInvoiceAmount: 0,
-      credInvoicePaidDate: null,
-      credInvoiceStatus: PAYMENT_STATUS.NOTPAID,
-    },
-    validationSchema: validationInvoiceUpdate,
-    onSubmit: () => {
-      null;
-    },
-  });
-
-  const formikPayInvoice = useFormik({
-    initialValues: {
       invoiceNo: '',
       invoiceAmount: 0,
       invoiceCreatedAt: new Date(),
     },
-    validationSchema: validationInvoicePayment,
+    validationSchema: validationSchema,
     onSubmit: () => {
       null;
     },
@@ -111,32 +84,14 @@ const CredPaymentsController = () => {
     setSearchTerm(event.target.value);
   };
 
-  const filteredData = invoices.filter((item) => {
-    const name = item.creditor.creditorName.toLowerCase();
-    const searchParamRegex = new RegExp(`${searchTerm.toLowerCase()}`, 'i');
-    return searchParamRegex.test(name);
-  });
-
-  const handleOpenCloseAddDialog = () => {
-    setIsOpenAdd(!isOpenAdd);
-
-    if (isOpenAdd) {
-      formikPayInvoice.resetForm();
-    }
-  };
-
-  const handleOpenCloseUpdateDialog = () => {
+  const handleOpenCloseUpdateDialog = (data) => {
     setIsOpenUpdate(!isOpenUpdate);
 
     if (!isOpenUpdate) {
       formik.setValues({
-        credInvoiceNo: selectedInvoice.credInvoiceNo,
-        credInvoiceAmount: selectedInvoice.credInvoiceAmount,
-        credInvoiceDate: new Date(selectedInvoice.credInvoiceDate),
-        credInvoicePaidDate: selectedInvoice.credInvoicePaidDate
-          ? new Date(selectedInvoice.credInvoicePaidDate)
-          : new Date(),
-        credInvoiceStatus: selectedInvoice.credInvoiceStatus,
+        invoiceNo: data.invoiceNo,
+        invoiceAmount: data.invoiceAmount,
+        invoiceCreatedAt: new Date(data.invoiceCreatedAt),
       });
     } else {
       formik.resetForm();
@@ -147,63 +102,29 @@ const CredPaymentsController = () => {
     setIsOpenDelete(!isOpenDelete);
   };
 
-  const handleSubmitAddPayment = async () => {
-    commonUtil.validateFormik(formikPayInvoice);
-
-    if (formikPayInvoice.isValid && formikPayInvoice.dirty) {
-      setIsLoadingAddPayment(true);
-
-      await backendAuthApi({
-        url: BACKEND_API.INVOICE_CREATE_CRED,
-        method: 'POST',
-        cancelToken: sourceToken.token,
-        data: {
-          credInvoiceId: selectedInvoice._id,
-          ...formikPayInvoice.values,
-        },
-      })
-        .then((res) => {
-          if (responseUtil.isResponseSuccess(res.data.responseCode)) {
-            handleOpenCloseAddDialog();
-            handleFetchPayments();
-          }
-          enqueueSnackbar(res.data.responseMessage, {
-            variant: responseUtil.findResponseType(res.data.responseCode),
-          });
-        })
-        .catch(() => {
-          setIsLoadingAddPayment(false);
-        })
-        .finally(() => {
-          setIsLoadingAddPayment(false);
-        });
-    } else {
-      enqueueSnackbar(SNACKBAR_MESSAGE.FILL_REQUIRED_FIELDS, { variant: SNACKBAR_VARIANT.WARNING });
-    }
-  };
-
   const handleSubmitUpdate = async () => {
     commonUtil.validateFormik(formik);
     if (formik.isValid && formik.dirty) {
       setIsLoadingUpdate(true);
 
       await backendAuthApi({
-        url: BACKEND_API.CREDITOR_UPDATE_INVOICE,
+        url: BACKEND_API.INVOICE_UPDATE_PAYMENT,
         method: 'PUT',
         cancelToken: sourceToken.token,
         data: {
-          credId: selectedInvoice._id,
+          invoiceId: selectedInvoice._id,
           ...formik.values,
         },
       })
         .then((res) => {
           if (responseUtil.isResponseSuccess(res.data.responseCode)) {
-            handleOpenCloseUpdateDialog(null);
             handleFetchPayments();
+            handleOpenCloseUpdateDialog(null);
+          } else {
+            enqueueSnackbar(res.data.responseMessage, {
+              variant: responseUtil.findResponseType(res.data.responseCode),
+            });
           }
-          enqueueSnackbar(res.data.responseMessage, {
-            variant: responseUtil.findResponseType(res.data.responseCode),
-          });
         })
         .catch(() => {
           setIsLoadingUpdate(false);
@@ -212,48 +133,47 @@ const CredPaymentsController = () => {
           setIsLoadingUpdate(false);
         });
     } else {
-      enqueueSnackbar(SNACKBAR_MESSAGE.FILL_REQUIRED_FIELDS, { variant: SNACKBAR_VARIANT.WARNING });
+      enqueueSnackbar(SNACKBAR_MESSAGE.FILL_REQUIRED_FIELDS, {
+        variant: SNACKBAR_VARIANT.WARNING,
+      });
     }
   };
 
   const handleSubmitDelete = async () => {
-    setIsLoadingDelete(true);
+    if (selectedInvoice) {
+      setIsLoadingDelete(true);
 
-    await backendAuthApi({
-      url: BACKEND_API.CREDITOR_DELETE_INVOICE + selectedInvoice._id,
-      method: 'DELETE',
-      cancelToken: sourceToken.token,
-    })
-      .then((res) => {
-        if (responseUtil.isResponseSuccess(res.data.responseCode)) {
-          handleOpenCloseDeleteDialog();
-          handleFetchPayments();
-        }
-        enqueueSnackbar(res.data.responseMessage, {
-          variant: responseUtil.findResponseType(res.data.responseCode),
+      await backendAuthApi({
+        url: BACKEND_API.INVOICE_DELETE_PAYMENT + selectedInvoice._id,
+        method: 'DELETE',
+        cancelToken: sourceToken.token,
+      })
+        .then((res) => {
+          if (responseUtil.isResponseSuccess(res.data.responseCode)) {
+            handleFetchPayments();
+            handleOpenCloseDeleteDialog();
+          }
+          enqueueSnackbar(res.data.responseMessage, {
+            variant: responseUtil.findResponseType(res.data.responseCode),
+          });
+        })
+        .catch(() => {
+          setIsLoadingDelete(false);
+        })
+        .finally(() => {
+          setIsLoadingDelete(false);
         });
-      })
-      .catch(() => {
-        setIsLoadingDelete(false);
-      })
-      .finally(() => {
-        setIsLoadingDelete(false);
-      });
+    }
   };
 
   const handleFetchTotalAmount = async () => {
-    const date = new Date(formikFilter.values.filteredDate);
-    const dayAfter = formikFilter.values.filteredDate
-      ? new Date(date.setDate(date.getDate() + 1))
-      : null;
-
     setIsLoadingTotal(true);
 
     await backendAuthApi({
       url: BACKEND_API.INVOICE_TOTAL_CRED_PAYMENTS,
       method: 'POST',
       cancelToken: sourceToken.token,
-      data: { filteredDate: dayAfter },
+      data: { filteredDate: formikFilter.values.filteredDate },
     })
       .then((res) => {
         if (responseUtil.isResponseSuccess(res.data.responseCode)) {
@@ -269,27 +189,23 @@ const CredPaymentsController = () => {
   };
 
   const handleFetchPayments = async () => {
-    const date = new Date(formikFilter.values.filteredDate);
-    const dayAfter = formikFilter.values.filteredDate
-      ? new Date(date.setDate(date.getDate() + 1))
-      : null;
-
     setIsLoading(true);
 
     await backendAuthApi({
-      url: BACKEND_API.CREDITORS_INVOICES,
-      method: 'POST',
+      url: BACKEND_API.INVOICE_CRED_PAYMENTS,
+      method: 'GET',
       cancelToken: sourceToken.token,
       params: {
         page: page,
         limit: rowsPerPage,
+        filteredDate: formikFilter.values.filteredDate,
+        customerName: searchTerm,
       },
-      data: { filteredDate: dayAfter },
     })
       .then((res) => {
         if (responseUtil.isResponseSuccess(res.data.responseCode)) {
           setInvoices(res.data.responseData.invoices);
-          setCount(res.data.responseData.count);
+          setCount(res.data.responseData.documentCount);
         }
       })
       .catch(() => {
@@ -305,7 +221,7 @@ const CredPaymentsController = () => {
     handleFetchTotalAmount();
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [formikFilter.values.filteredDate, page, rowsPerPage]);
+  }, [formikFilter.values.filteredDate, page, rowsPerPage, searchTerm]);
 
   return (
     <CredPaymentsView
@@ -317,19 +233,13 @@ const CredPaymentsController = () => {
       setSelectedInvoice={setSelectedInvoice}
       searchTerm={searchTerm}
       handleSearchInputChange={handleSearchInputChange}
-      filteredData={filteredData}
       formik={formik}
-      formikPayInvoice={formikPayInvoice}
-      isOpenAdd={isOpenAdd}
       isOpenUpdate={isOpenUpdate}
       isOpenDelete={isOpenDelete}
-      isLoadingAddPayment={isLoadingAddPayment}
       isLoadingUpdate={isLoadingUpdate}
       isLoadingDelete={isLoadingDelete}
-      handleOpenCloseAddDialog={handleOpenCloseAddDialog}
       handleOpenCloseUpdateDialog={handleOpenCloseUpdateDialog}
       handleOpenCloseDeleteDialog={handleOpenCloseDeleteDialog}
-      handleSubmitAddPayment={handleSubmitAddPayment}
       handleSubmitUpdate={handleSubmitUpdate}
       handleSubmitDelete={handleSubmitDelete}
       handleFetchPayments={handleFetchPayments}
